refactor to lift the library search code closer to where it\'s used
