+++ src/renderer/screens/Editor/Sidebar/LayoutSharing.js
@@ -10,10 +10,7 @@ import Drawer from "@mui/material/Drawer";
 import IconButton from "@mui/material/IconButton";
 import Typography from "@mui/material/Typography";
 import ConfirmationDialog from "@renderer/components/ConfirmationDialog";
-import { getStaticPath } from "@renderer/config";
-import fs from "fs";
 import { t } from "i18next";
-import path from "path";
 import React from "react";
 import { BackupImport } from "./LayoutSharing/BackupImport";
 import { ExportToFile } from "./LayoutSharing/ExportToFile";
@@ -26,38 +23,11 @@ export default class LayoutSharing extends React.Component {
   constructor(props) {
     super(props);
 
-    const { vendor, product } = focus.focusDeviceDescriptor.info;
-    const cVendor = vendor.replace("/", "");
-    const cProduct = product.replace("/", "");
-    const layoutDirPath = path.join(
-      getStaticPath(),
-      cVendor,
-      cProduct,
-      "layouts"
-    );
-
-    try {
-      const layouts = fs
-        .readdirSync(layoutDirPath, {
-          encoding: "utf-8",
-        })
-        .map((name) => path.basename(name, ".json"))
-        .sort();
-
-      this.state = {
-        importConfirmOpen: false,
-        layout: {},
-        layoutName: null,
-        library: layouts,
-      };
-    } catch (_) {
-      this.state = {
-        importConfirmOpen: false,
-        layout: {},
-        layoutName: null,
-        library: [],
-      };
-    }
+    this.state = {
+      importConfirmOpen: false,
+      layout: {},
+      layoutName: null,
+    };
   }
 
   setLayout = (layoutName, layout) => {
@@ -97,7 +67,7 @@ export default class LayoutSharing extends React.Component {
 
   render() {
     const { open, onClose, theme, keymap, colormap, ...others } = this.props;
-    const { layout, layoutName, library } = this.state;
+    const { layout, layoutName } = this.state;
     const sidebarWidth = 300;
 
     const Keymap = focus.focusDeviceDescriptor.components.keymap;
@@ -106,14 +76,6 @@ export default class LayoutSharing extends React.Component {
     const previewColormap = layout.colormaps
       ? layout.colormaps[0]
       : colormap.colorMap[0];
-    const keymapWidget = (
-      <Keymap
-        keymap={previewLayout}
-        palette={palette}
-        colormap={previewColormap}
-        theme={theme}
-      />
-    );
 
     return (
       <Dialog open={open} onClose={onClose} fullScreen>
@@ -148,7 +110,6 @@ export default class LayoutSharing extends React.Component {
             <Box sx={{ overflow: "auto", padding: 3 }}>
               <LibraryImport
                 setLayout={this.setLayout}
-                library={library}
                 layoutName={layoutName}
                 {...others}
               />
@@ -187,7 +148,12 @@ export default class LayoutSharing extends React.Component {
               width: `calc(100% - ${sidebarWidth}px)`,
             }}
           >
-            {keymapWidget}
+            <Keymap
+              keymap={previewLayout}
+              palette={palette}
+              colormap={previewColormap}
+              theme={theme}
+            />{" "}
           </Box>
         </DialogContent>
       </Dialog>
